# spec/api-openapi.json
{
  "components": {
    "schemas": {
...
      "AuditRunOut": {
        "properties": {
          "audited_up_to_chapter": {
            "title": "Audited Up To Chapter",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "findings": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Findings",
            "type": "array"
          },
          "sampled_characters": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Sampled Characters",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          },
          "window_end": {
            "title": "Window End",
            "type": "integer"
          },
          "window_start": {
            "title": "Window Start",
            "type": "integer"
          }
        },
        "required": [
          "window_start",
          "window_end",
          "audited_up_to_chapter",
          "status"
        ],
        "title": "AuditRunOut",
        "type": "object"
      },
...
      "GlobalAuditDetailOut": {
        "description": "详情：列表项 + 抽样角色 + findings 明细（覆盖 Summary 的 int 计数为 list，同前端 Omit 模式）。",
        "properties": {
          "audited_up_to_chapter": {
            "title": "Audited Up To Chapter",
            "type": "integer"
          },
          "bridge": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bridge"
          },
          "chapters": {
            "title": "Chapters",
            "type": "integer"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "findings": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Findings",
            "type": "array"
          },
          "report_id": {
            "title": "Report Id",
            "type": "string"
          },
          "sampled": {
            "title": "Sampled",
            "type": "integer"
          },
          "sampled_characters": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Sampled Characters",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "style": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Style"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          },
          "trigger": {
            "title": "Trigger",
            "type": "string"
          },
          "volume": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          },
          "window_end": {
            "title": "Window End",
            "type": "integer"
          },
          "window_start": {
            "title": "Window Start",
            "type": "integer"
          }
        },
        "required": [
          "report_id",
          "window_start",
          "window_end",
          "audited_up_to_chapter",
          "trigger",
          "status",
          "sampled",
          "chapters"
        ],
        "title": "GlobalAuditDetailOut",
        "type": "object"
      },
      "GlobalAuditSummaryOut": {
        "properties": {
          "audited_up_to_chapter": {
            "title": "Audited Up To Chapter",
            "type": "integer"
          },
          "bridge": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bridge"
          },
          "chapters": {
            "title": "Chapters",
            "type": "integer"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "findings": {
            "title": "Findings",
            "type": "integer"
          },
          "report_id": {
            "title": "Report Id",
            "type": "string"
          },
          "sampled": {
            "title": "Sampled",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "style": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Style"
          },
          "trigger": {
            "title": "Trigger",
            "type": "string"
          },
          "volume": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          },
          "window_end": {
            "title": "Window End",
            "type": "integer"
          },
          "window_start": {
            "title": "Window Start",
            "type": "integer"
          }
        },
        "required": [
          "report_id",
          "window_start",
          "window_end",
          "audited_up_to_chapter",
          "trigger",
          "status",
          "sampled",
          "findings",
          "chapters"
        ],
        "title": "GlobalAuditSummaryOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LessonActionOut": {
        "properties": {
          "lesson_id": {
            "title": "Lesson Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "lesson_id",
          "status"
        ],
        "title": "LessonActionOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WritingLessonOut": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Evidence",
            "type": "array"
          },
          "last_recurrence_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Recurrence At"
          },
          "lesson_id": {
            "title": "Lesson Id",
            "type": "string"
          },
          "lesson_type": {
            "title": "Lesson Type",
            "type": "string"
          },
          "recurrence_count": {
            "title": "Recurrence Count",
            "type": "integer"
          },
          "source_chapter": {
            "title": "Source Chapter",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "lesson_id",
          "category",
          "lesson_type",
          "content",
          "confidence",
          "source_chapter",
          "status",
          "recurrence_count"
        ],
        "title": "WritingLessonOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Myink 内部 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/v1/projects/{project_id}/global-audit": {
      "get": {
        "description": "审计报告列表（最新在前，limit ≤ 20）：窗口/状态/计数，前端审计视图导航。",
        "operationId": "list_global_audits_internal_v1_projects__project_id__global_audit_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Myink-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Myink-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GlobalAuditSummaryOut"
                  },
                  "title": "Response List Global Audits Internal V1 Projects  Project Id  Global Audit Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Global Audits",
        "tags": [
          "global-audit"
        ]
      },
      "post": {
        "description": "手动触发全局审计：全部未审计章窗口（不做 K 门槛），返回审计报告。\n\n无已写章节 → 400；LLM/解析失败 → 502（报告已落库为 failed 并推进 marker）。",
        "operationId": "trigger_global_audit_internal_v1_projects__project_id__global_audit_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Myink-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Myink-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditRunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trigger Global Audit",
        "tags": [
          "global-audit"
        ]
      }
    },
    "/internal/v1/projects/{project_id}/global-audit/{report_id}": {
      "get": {
        "description": "单报告详情：findings 明细 + 对照的卷/阶段 + 计数。",
        "operationId": "get_global_audit_internal_v1_projects__project_id__global_audit__report_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Myink-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Myink-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalAuditDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Global Audit",
        "tags": [
          "global-audit"
        ]
      }
    },
...
    "/internal/v1/projects/{project_id}/lessons": {
      "get": {
        "description": "写作经验列表（reflexion 复盘，按来源章排序；status 可选过滤）。",
        "operationId": "list_lessons_internal_v1_projects__project_id__lessons_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "header",
            "name": "X-Myink-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Myink-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WritingLessonOut"
                  },
                  "title": "Response List Lessons Internal V1 Projects  Project Id  Lessons Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Lessons",
        "tags": [
          "lessons"
        ]
      }
    },
    "/internal/v1/projects/{project_id}/lessons/{lesson_id}/confirm": {
      "post": {
        "description": "确认经验生效（proposed→active，后续章节注入遵守）。幂等：非 proposed 返回 409。",
        "operationId": "confirm_lesson_internal_v1_projects__project_id__lessons__lesson_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "lesson_id",
            "required": true,
            "schema": {
              "title": "Lesson Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Myink-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Myink-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonActionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Lesson",
        "tags": [
          "lessons"
        ]
      }
    },
    "/internal/v1/projects/{project_id}/lessons/{lesson_id}/reject": {
      "post": {
        "description": "拒绝经验（proposed→rejected，幻觉/无益经验清理）。幂等：非 proposed 返回 409。",
        "operationId": "reject_lesson_internal_v1_projects__project_id__lessons__lesson_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "lesson_id",
            "required": true,
            "schema": {
              "title": "Lesson Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Myink-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Myink-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonActionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Lesson",
        "tags": [
          "lessons"
        ]
      }
    },
...
  }
}